BI-10397: Refactor duplicated constants into interface.
