+++ document-generator-company-report/src/main/java/uk/gov/companieshouse/document/generator/company/report/service/ChargesService.java
@@ -9,14 +9,15 @@ import uk.gov.companieshouse.api.handler.exception.URIValidationException;
 import uk.gov.companieshouse.api.model.charges.ChargesApi;
 import uk.gov.companieshouse.document.generator.company.report.exception.ServiceException;
 
+import static uk.gov.companieshouse.document.generator.company.report.service.PageRetrieverClientConstants.ITEMS_PER_PAGE_KEY;
+import static uk.gov.companieshouse.document.generator.company.report.service.PageRetrieverClientConstants.ITEMS_PER_PAGE_VALUE;
+import static uk.gov.companieshouse.document.generator.company.report.service.PageRetrieverClientConstants.START_INDEX_KEY;
+
 @Service
 public class ChargesService implements PageRetrieverClient<ChargesApi> {
 
-    public static final String ITEMS_PER_PAGE_KEY = "items_per_page";
-    public static final String START_INDEX_KEY = "start_index";
     private static final UriTemplate GET_CHARGES_URI =
         new UriTemplate("/company/{companyNumber}/charges");
-    private static final int ITEMS_PER_PAGE_VALUE = 100;
 
     private final CompanyReportApiClientService companyReportApiClientService;
     private final PageRetrieverService<ChargesApi> pageRetrieverService;
