+++ document-generator-company-report/src/main/java/uk/gov/companieshouse/document/generator/company/report/service/PageRetrieverClientConstants.java
@@ -1,0 +1,10 @@
+package uk.gov.companieshouse.document.generator.company.report.service;
+
+/**
+ * Constants that are useful for {@link PageRetrieverClient} implementations.
+ */
+interface PageRetrieverClientConstants {
+    String ITEMS_PER_PAGE_KEY = "items_per_page";
+    String START_INDEX_KEY = "start_index";
+    int ITEMS_PER_PAGE_VALUE = 100;
+}
