+++ document-generator-company-report/src/test/java/uk/gov/companieshouse/document/generator/company/report/service/ChargesServiceTest.java
@@ -28,8 +28,8 @@ import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static uk.gov.companieshouse.document.generator.company.report.service.ChargesService.ITEMS_PER_PAGE_KEY;
-import static uk.gov.companieshouse.document.generator.company.report.service.ChargesService.START_INDEX_KEY;
+import static uk.gov.companieshouse.document.generator.company.report.service.PageRetrieverClientConstants.ITEMS_PER_PAGE_KEY;
+import static uk.gov.companieshouse.document.generator.company.report.service.PageRetrieverClientConstants.START_INDEX_KEY;
 
 @ExtendWith(MockitoExtension.class)
 public class ChargesServiceTest {
